+++ tests/net/iface/src/Makefile
@@ -1,4 +1,4 @@
-obj-y = main.o
+obj-y = main.o test_iface_context.o
 ccflags-y += -I${ZEPHYR_BASE}/subsys/net/ip
 
 include $(ZEPHYR_BASE)/tests/Makefile.test
